test: move disabled wildcard URL mappings to separate test app
