+++ grails-test-examples/app5/build.gradle
@@ -16,27 +16,26 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package functionaltests.urlmappings
 
-import grails.compiler.GrailsCompileStatic
-import grails.converters.JSON
-
-@GrailsCompileStatic
-class WildcardValidationController {
-
-    static responseFormats = ['json']
-
-    def show() {
-        def payload = [
-            controller: 'wildcardValidation',
-            action: 'show'
-        ]
-        if (params.id) {
-            payload.id = params.id
-        }
-        render(payload as JSON)
-    }
+plugins {
+    id 'org.apache.grails.buildsrc.properties'
+    id 'org.apache.grails.buildsrc.compile'
+    id 'org.apache.grails.gradle.grails-web'
 }
 
+version = '0.1'
+group = 'app5'
 
+dependencies {
 
+    implementation platform(project(':grails-bom'))
+    implementation 'org.apache.grails:grails-dependencies-starter-web'
+
+    integrationTestImplementation 'org.apache.grails:grails-testing-support-http-client'
+    integrationTestImplementation 'org.apache.grails:grails-testing-support-web'
+}
+
+apply {
+    from rootProject.layout.projectDirectory.file('gradle/functional-test-config.gradle')
+    from rootProject.layout.projectDirectory.file('gradle/grails-extension-gradle-config.gradle')
+}
