+++ grails-test-examples/app5/grails-app/controllers/app5/UrlMappings.groovy
@@ -1,0 +1,35 @@
+/*
+ *  Licensed to the Apache Software Foundation (ASF) under one
+ *  or more contributor license agreements.  See the NOTICE file
+ *  distributed with this work for additional information
+ *  regarding copyright ownership.  The ASF licenses this file
+ *  to you under the Apache License, Version 2.0 (the
+ *  "License"); you may not use this file except in compliance
+ *  with the License.  You may obtain a copy of the License at
+ *
+ *    https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing,
+ *  software distributed under the License is distributed on an
+ *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ *  KIND, either express or implied.  See the License for the
+ *  specific language governing permissions and limitations
+ *  under the License.
+ */
+package app5
+
+class UrlMappings {
+
+    static mappings = {
+
+        '/wildcard-disabled/show-action'(controller: 'wildcardValidation', action: 'show')
+        '/wildcard-disabled/$action'(controller: 'wildcardValidation')
+        '/wildcard-disabled/$id'(controller: 'wildcardValidation', action: 'show')
+
+        "/$controller/$action?/$id?(.$format)?"{
+            constraints {
+                // apply constraints here
+            }
+        }
+    }
+}
