+++ settings.gradle
@@ -366,6 +366,7 @@ include(
         'grails-test-examples-app2',
         'grails-test-examples-app3',
         'grails-test-examples-app4',
+        'grails-test-examples-app5',
         'grails-test-examples-external-configuration',
         'grails-test-examples-exploded',
         'grails-test-examples-datasources',
@@ -404,6 +405,7 @@ project(':grails-test-examples-app1').projectDir = file('grails-test-examples/ap
 project(':grails-test-examples-app2').projectDir = file('grails-test-examples/app2')
 project(':grails-test-examples-app3').projectDir = file('grails-test-examples/app3')
 project(':grails-test-examples-app4').projectDir = file('grails-test-examples/app4')
+project(':grails-test-examples-app5').projectDir = file('grails-test-examples/app5')
 project(':grails-test-examples-external-configuration').projectDir = file('grails-test-examples/external-configuration')
 project(':grails-test-examples-exploded').projectDir = file('grails-test-examples/exploded')
 project(':grails-test-examples-datasources').projectDir = file('grails-test-examples/datasources')
